+++ src/PrintOrder/utils.js
@@ -1,18 +1,11 @@
-import { pick } from 'lodash';
+import pick from 'lodash/pick';
 
 import {
+  fetchTenantAddressesByIds,
   VENDORS_API,
-  CONFIG_API,
 } from '@folio/stripes-acq-components';
 
-import {
-  CONFIG_ADDRESSES,
-  MODULE_TENANT,
-} from '../components/Utils/const';
-import {
-  getAddresses,
-  getRecordMap,
-} from '../common/utils';
+import { getRecordMap } from '../common/utils';
 
 export const getPrintPageStyles = () => `
   @page {
@@ -48,14 +41,10 @@ export const buildAddressString = (address = {}) => (
 export const getOrderPrintData = async (ky, order = {}) => {
   const vendor = await ky.get(`${VENDORS_API}/${order.vendor}`).json().catch(() => ({}));
   const addressIds = [...new Set([order.billTo, order.shipTo])].filter(Boolean);
-  const subQuery = addressIds.map(id => `id==${id}`).join(' or ');
-  const searchParams = {
-    query: `(module=${MODULE_TENANT} and configName=${CONFIG_ADDRESSES} and (${subQuery}))`,
-  };
-  const addressesResp = addressIds.length
-    ? await ky.get(CONFIG_API, { searchParams }).json().catch(() => ({}))
-    : {};
-  const addresses = getAddresses(addressesResp?.configs || []);
+
+  const addresses = addressIds.length
+    ? await fetchTenantAddressesByIds(ky)(addressIds).then((res) => res.addresses).catch(() => [])
+    : [];
   const addressMap = getRecordMap(addresses);
 
   return ({